Fix display of error messages on paper-autocomplete
The 'isTouched' property of the paper-input field - used as trigger
component for paper-autocomplete - is not correctly updated to 'true' on
focus-out of the autocomplete. As a consequence the error messages are
correctly set on the paper-input component, but not displayed in the
template.
We workaround this by permanently setting the isTouched property of the
autocomplete trigger component to 'true'.