Fix selection of values in ember-power-select
Due to a change in ember-data the Proxy object of a belongsTo relation
doesn't always update anymore - unless set to a null value - but
proxy.content does. As a consequence, ember-power-select doesn't get
updated anymore on changes of the selected value if the proxy object is
passed as value. To fix this, we now pass proxy.content as value to ember-power-select. Note that
this makes the component unusable for non-proxy-object values.

Follow-up of the issue can be found in https://github.com/cibernox/ember-power-select/issues/1147.