Make selects independent of ObjectProxy values
Select component must have the same interface independent whether the
passed value is a plain object or an ObjectProxy with a content
attribute.
Due to a change in the behaviour of ObjectProxies in a belongsTo
relationship in Ember data the select components didn't get updated
correctly if an ObjectProxy is passed as value. This commit adds a
workaround so this behaviour is not exposed to the consumer of the
select components.

See also https://github.com/cibernox/ember-power-select/issues/1147