Add support for offers created in Access
Display total amount of offers created in Access. Whether an offer is
created in Access is determined based on the 'amount' field. If it's
set, the offer has been created in Access.
If an order has already been created, the content of the offer cannot be
updated anymore. If an order has not yet been created, the user can add
an offerline. As from that moment, the offer will no longer be mastered
by Access.