Don't unlink phases manually on destroy of last phase
Calling destroyRecord automatically removes the link between 2 resources
in the frontend. The backend will also be updated correctly by the
delete request.

Unlinking manually and saving the previous phase before removing the
final phase leads to additional complexity because in the update request
to the backend because .Net Core cannot make a distinction between null
and undefined relations.