Add CSS transition on opening and closing of menu
To ensure the CSS transition on close is visible, the flag to
determine whether the menu is opened must be split in a flag for the
wrapping div and a flag for the inner content of the menu.

The flag for the inner content changes immediately on open/close but the wrapping div
must stay visibled as long as the leave transition takes, 300ms in
this case, for the animation to be visible.