Make amount a required field of a calculation line
Amount must be required on each calculation line. However, to ensure
the first calculation-line automatically created on offerline creation
appears in the list, it must be persisted in the backend
already. Therefore validation is not executed before saving the first
calculation-line. For additional calculation-lines the validation  is
applied before saving. Hence, if no amount is filled in, the
calculation-line will not be persisted.

Missing amounts are indicated with a red border on the
input-field. Invalid calculation-lines are automatically opened in edit-mode.